[bugfix] ffmpeg screen recoder lost frame on mac 10.15.
why:Video and audio are collected in the same thread, and the process of collecting and processing pictures takes too long, resulting in some audio being covered without collection.
how：When the audio ready signal is received, the audio is immediately collected into the cache, rather than waiting for the upper layer to collect.